Have test coverage at 100% for a lot of the classes. Still more left though.

+++ EfficientRealtimeJava3D/src/Matrix4f.java
@@ -358,7 +358,7 @@ public class Matrix4f implements java.io.Serializable
 		array[15] = this.m33;
 	}
 
-	public double get(int row, int column)
+	public float get(int row, int column)
 	{
 		switch (row)
 		{
